chore: switch to .native entrypoints, use web as default entrypoint
