+++ src/specs/index.ts
@@ -1,27 +1,1 @@
-import './TurboUnistyles'
-
-import { Runtime } from './UnistylesRuntime'
-import { StyleSheet } from './StyleSheet'
-import type { UnistylesStyleSheet } from './StyleSheet'
-import { UnistylesShadowRegistry } from './ShadowRegistry'
-import type { UnistylesMiniRuntime } from './UnistylesRuntime'
-import { StatusBarStyle, ColorScheme, Orientation } from './types'
-import { UnistyleDependency } from './NativePlatform'
-
-export {
-    StatusBarStyle,
-    ColorScheme,
-    Orientation,
-    UnistyleDependency
-}
-
-export {
-    UnistylesShadowRegistry,
-    Runtime as UnistylesRuntime,
-    StyleSheet,
-}
-
-export type {
-    UnistylesMiniRuntime,
-    UnistylesStyleSheet
-}
+export * from '../web'
