--- src/specs/index.web.ts
@@ -1,1 +1,0 @@
-export * from '../web'
